Calling (paredit-mode) etc. from elisp code always enables, not toggles, the mode.
`paredit-mode' with no arguments only toggles the mode when called
interactively. From Lisp, ommitted or `nil' argument enables the mode -- this
"makes it easy to enable the minor mode in a major mode hook, for example".

See: (info "(elisp) Defining Minor Modes")

Third parties using `esk-turn-on-paredit', `esk-turn-on-whitespace', or
`esk-turn-on-idle-highlight-mode' should simply replace
  (add-hook 'xyz 'esk-turn-on-paredit)
with
  (add-hook 'xyz 'paredit-mode)